net: Move RHCOS9_WORKER_LABEL to higher lib To be used in net modules
